@@ -1,10 +1,10 @@
 import React from "react";
 
-export function InputLayout(props: any) {
-    return <>
-        <div className="input-layout">
-            <label>{props.label}</label>
-            {props.children}
-        </div>
-    </>;
+export function InputLayout({ label, children }: any) {
+	return (
+		<div className="input-layout">
+			<label>{label}</label>
+			{children}
+		</div>
+	);
 }
